+++ sync-for-payables/src/funcs/suppliersCreate.ts
@@ -154,7 +154,7 @@ export async function suppliersCreate(
     | RequestTimeoutError
     | ConnectionError
   >(
-    M.json(200, shared.Supplier$inboundSchema),
+    M.json(201, shared.Supplier$inboundSchema),
     M.jsonErr(
       [400, 401, 402, 403, 404, 429, 500, 503],
       errors.ErrorMessage$inboundSchema,